requirements.txt
pywin32>=306
Pillow>=10.0.0
pystray>=0.19.4
pyinstaller>=5.0